change: Split ReportCommand and ReportsCommand into two modules
