+++ src/gmp/gmp.ts
@@ -43,7 +43,8 @@ import FeedStatusCommand from 'gmp/commands/feedstatus';
 import LoginCommand from 'gmp/commands/login';
 import PerformanceCommand from 'gmp/commands/performance';
 import {PortListCommand, PortListsCommand} from 'gmp/commands/portlists';
-import {ReportCommand, ReportsCommand} from 'gmp/commands/reports';
+import ReportCommand from 'gmp/commands/report';
+import ReportsCommand from 'gmp/commands/reports';
 import RoleCommand from 'gmp/commands/role';
 import RolesCommand from 'gmp/commands/roles';
 import TaskCommand from 'gmp/commands/task';
